SDFSOF-64 - Fixed review comments. Commented out integration tests
