+++ platform/src/main/java/org/stellar/anchor/platform/action/ActionHandler.java
@@ -1,5 +1,6 @@
 package org.stellar.anchor.platform.action;
 
+import static java.util.Comparator.comparing;
 import static java.util.stream.Collectors.joining;
 import static java.util.stream.Collectors.toList;
 import static org.stellar.anchor.api.sep.AssetInfo.NATIVE_ASSET_CODE;
@@ -7,13 +8,17 @@ import static org.stellar.anchor.api.sep.SepTransactionStatus.COMPLETED;
 import static org.stellar.anchor.api.sep.SepTransactionStatus.ERROR;
 import static org.stellar.anchor.api.sep.SepTransactionStatus.EXPIRED;
 import static org.stellar.anchor.api.sep.SepTransactionStatus.REFUNDED;
+import static org.stellar.anchor.platform.service.PaymentOperationToEventListener.parsePaymentTime;
 import static org.stellar.anchor.platform.utils.TransactionHelper.toGetTransactionResponse;
+import static org.stellar.anchor.util.Log.error;
 import static org.stellar.anchor.util.Log.errorEx;
 import static org.stellar.anchor.util.MathHelper.decimal;
 
+import java.io.IOException;
 import java.time.Instant;
 import java.util.Arrays;
 import java.util.List;
+import java.util.Objects;
 import java.util.Set;
 import javax.validation.ConstraintViolation;
 import javax.validation.Validator;
@@ -21,6 +26,7 @@ import org.apache.commons.collections.CollectionUtils;
 import org.springframework.transaction.annotation.Transactional;
 import org.stellar.anchor.api.exception.AnchorException;
 import org.stellar.anchor.api.exception.BadRequestException;
+import org.stellar.anchor.api.exception.SepException;
 import org.stellar.anchor.api.exception.rpc.InvalidParamsException;
 import org.stellar.anchor.api.exception.rpc.InvalidRequestException;
 import org.stellar.anchor.api.platform.GetTransactionResponse;
@@ -29,11 +35,15 @@ import org.stellar.anchor.api.rpc.action.AmountRequest;
 import org.stellar.anchor.api.rpc.action.RpcActionParamsRequest;
 import org.stellar.anchor.api.sep.AssetInfo;
 import org.stellar.anchor.api.sep.SepTransactionStatus;
+import org.stellar.anchor.api.shared.Amount;
+import org.stellar.anchor.api.shared.StellarPayment;
+import org.stellar.anchor.api.shared.StellarTransaction;
 import org.stellar.anchor.asset.AssetService;
 import org.stellar.anchor.horizon.Horizon;
 import org.stellar.anchor.platform.data.JdbcSep24Transaction;
 import org.stellar.anchor.platform.data.JdbcSep31Transaction;
 import org.stellar.anchor.platform.data.JdbcSepTransaction;
+import org.stellar.anchor.platform.observer.ObservedPayment;
 import org.stellar.anchor.sep24.Sep24TransactionStore;
 import org.stellar.anchor.sep31.Sep31Transaction;
 import org.stellar.anchor.sep31.Sep31TransactionStore;
@@ -41,13 +51,15 @@ import org.stellar.anchor.util.SepHelper;
 import org.stellar.anchor.util.StringHelper;
 import org.stellar.sdk.AssetTypeCreditAlphaNum;
 import org.stellar.sdk.responses.AccountResponse;
+import org.stellar.sdk.responses.operations.PathPaymentBaseOperationResponse;
+import org.stellar.sdk.responses.operations.PaymentOperationResponse;
 
 public abstract class ActionHandler<T extends RpcActionParamsRequest> {
 
   protected final Sep24TransactionStore txn24Store;
   protected final Sep31TransactionStore txn31Store;
   private final Validator validator;
-  protected final Horizon horizon;
+  private final Horizon horizon;
   protected final AssetService assetService;
   private final List<AssetInfo> assets;
 
@@ -203,6 +215,81 @@ public abstract class ActionHandler<T extends RpcActionParamsRequest> {
           String.format("Unable to check trust for account[%s] and asset[%s]", account, asset), e);
       return false;
     }
+  }
+
+  protected void addStellarTransaction(JdbcSepTransaction txn, String stellarTransactionId) {
+    try {
+      List<ObservedPayment> payments = getObservedPayments(stellarTransactionId);
+
+      if (!payments.isEmpty()) {
+        ObservedPayment firstPayment = payments.get(0);
+        ObservedPayment lastPayment = payments.get(payments.size() - 1);
+        JdbcSep24Transaction txn24 = (JdbcSep24Transaction) txn;
+
+        StellarTransaction stellarTransaction =
+            StellarTransaction.builder()
+                .id(stellarTransactionId)
+                .createdAt(parsePaymentTime(firstPayment.getCreatedAt()))
+                .memo(txn24.getMemo())
+                .memoType(txn24.getMemoType())
+                .envelope(firstPayment.getTransactionEnvelope())
+                .payments(
+                    payments.stream()
+                        .map(
+                            payment ->
+                                StellarPayment.builder()
+                                    .id(payment.getId())
+                                    .amount(new Amount(payment.getAmount(), payment.getAssetName()))
+                                    .sourceAccount(payment.getFrom())
+                                    .destinationAccount(payment.getTo())
+                                    .paymentType(
+                                        payment.getType() == ObservedPayment.Type.PAYMENT
+                                            ? StellarPayment.Type.PAYMENT
+                                            : StellarPayment.Type.PATH_PAYMENT)
+                                    .build())
+                        .collect(toList()))
+                .build();
+
+        if (txn.getStellarTransactions() == null) {
+          txn.setStellarTransactions(List.of(stellarTransaction));
+        } else {
+          txn.getStellarTransactions().add(stellarTransaction);
+        }
+        txn.setTransferReceivedAt(parsePaymentTime(lastPayment.getCreatedAt()));
+      }
+    } catch (Exception ex) {
+      errorEx("Failed to retrieve stellar transactions", ex);
+    }
+  }
+
+  private List<ObservedPayment> getObservedPayments(String stellarTransactionId)
+      throws IOException {
+    return horizon
+        .getServer()
+        .payments()
+        .includeTransactions(true)
+        .forTransaction(stellarTransactionId)
+        .execute()
+        .getRecords()
+        .stream()
+        .map(
+            operation -> {
+              try {
+                if (operation instanceof PaymentOperationResponse) {
+                  return ObservedPayment.fromPaymentOperationResponse(
+                      (PaymentOperationResponse) operation);
+                } else if (operation instanceof PathPaymentBaseOperationResponse) {
+                  return ObservedPayment.fromPathPaymentOperationResponse(
+                      (PathPaymentBaseOperationResponse) operation);
+                }
+              } catch (SepException e) {
+                error("Failed to parse operation response");
+              }
+              return null;
+            })
+        .filter(Objects::nonNull)
+        .sorted(comparing(ObservedPayment::getCreatedAt))
+        .collect(toList());
   }
 
   private void updateTransaction(JdbcSepTransaction txn, RpcActionParamsRequest request)
